refactor(EMA): move eds-specific code to subfolder
