+++ src/sophys/cli/extensions/ema/eds/open_plugin_page.sh
